refactor(decompose): split packages/libraries/graph-tools/src/commands/capture/run.ts into folder
